aws/amazon_strands/setup-agent/my_agent/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.28.0